pyproject.toml
[build-system]
    build-backend = "setuptools.build_meta"
    requires      = ["setuptools>=64"]

[project]
    authors = [{ name = "Espressif Systems" }]
    classifiers = [
        "Intended Audience :: Developers",
        "License :: OSI Approved :: Apache Software License",
        "Natural Language :: English",
        "Programming Language :: Python :: 3",
    ]
    description = "Espressif GitHub Action for creating recursive source zips with submodules on tag push"
    dynamic = ["version"]
    license = { text = "Apache 2.0" }
    name = "release-zips-action"
    readme = "README.md"
    requires-python = ">=3.11"
    urls.Changelog = "https://github.com/espressif/release-zips-action/tags/blob/master/CHANGELOG.md"
    urls.Homepage = "https://github.com/espressif/release-zips-action/tags"
    urls.Source = "https://github.com/espressif/release-zips-action/tags"
    urls.Tracker = "https://github.com/espressif/release-zips-action/tags/issues"

    dependencies              = ["PyGithub~=2.6.1"]
    optional-dependencies.dev = ["czespressif", "just-bin~=1.26.0", "pip-tools~=7.3", "pre-commit>=3.3"]

[tool]
    [tool.setuptools]
        dynamic.version       = { attr = "release_zips.__init__.__version__" }
        packages.find.exclude = ["tests"]
        packages.find.include = ["release_zips"]

    [tool.codespell]
        skip = "tests"

    [tool.ruff]
        extend-exclude                 = ["tests/"]
        format.quote-style             = "single"
        line-length                    = 120
        lint.exclude                   = ["tests/**/*"]
        lint.ignore                    = ["S603", "S607"]
        lint.isort.force-single-line   = true
        lint.isort.lines-between-types = 1
        lint.select                    = ["A", "B", "C4", "C901", "E", "F", "I", "N", "PL", "S", "UP", "W"]

    [tool.mypy]
        disallow_incomplete_defs = true
        disallow_untyped_defs    = false
        exclude                  = ["^__init__.py$", "^build/.*", "^dist/.*", "^docs/.*", "^tests/.*"]
        ignore_missing_imports   = true
        packages                 = "release_zips"
        pretty                   = true
        python_version           = "3.11"
        warn_return_any          = true

    [tool.vulture]
        make_whitelist = true
        min_confidence = 80
        paths          = ["release_zips"]
        sort_by_size   = true
        verbose        = true

    [tool.commitizen]
        annotated_tag              = true
        bump_message               = 'change(bump): [skip-ci] release $current_version → $new_version'
        changelog_merge_prerelease = true
        changelog_start_rev        = "v0.1.0"
        name                       = "czespressif"
        tag_format                 = "v$version"
        update_changelog_on_bump   = true
        version                    = "0.1.0"
        version_files              = ["release_zips/__init__.py"]
        version_provider           = "commitizen"

    [tool.yamlfix]
        comments_min_spaces_from_content = 1
        line_length                      = 200
        preserve_quotes                  = true
        section_whitelines               = 1
        sequence_style                   = "keep_style"
        whitelines                       = 1

    [tool.pytest]
        ini_options.addopts   = "-s --log-cli-level DEBUG --cov=czespressif --cov-report=term"
        ini_options.markers   = ["settings: mark test to use custom czespressif settings from pyproject.toml"]
        ini_options.testpaths = "tests"

    [tool.coverage]
        run.branch   = true
        run.omit     = ["*/tests/*", "__*__.py"]
        run.parallel = true
        run.source   = ["release_zips"]

    [tool.pip-tools]
        allow-unsafe     = true
        annotation-style = "line"
        strip-extras     = true
